fix: add check for deleting local file

+++ lib/desktop_screens_new/my_files_screen/widgets/category_screen.dart
@@ -421,7 +421,7 @@ class _CategoryScreenState extends State<CategoryScreen> {
 
                                 File localFile = File(filePath);
                                 bool fileExists = await localFile.exists();
-                                if (fileExists) {
+                                if (fileExists && res) {
                                   localFile.deleteSync();
                                 }
